+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ Changed
 -------
 - Change ``variant-annotation`` process
   to skip downstream/upstream and intergenic variants
+- Change clinical diagnosis and annotation fields type to text
 
 Fixed
 -----